move around historical graphcs
