+++ software/tracksight/frontend/src/app/visualize/timestamp_picker.tsx
